Give List, Map, ObjectDetail and PivotTable their own definition file
