+++ frontend/src/metabase/visualizations/visualizations/ObjectDetail/index.ts
@@ -1,0 +1,1 @@
+export { ObjectDetail } from "./ObjectDetail";
