+++ frontend/src/metabase/visualizations/visualizations/PivotTable/PivotTable.tsx
@@ -1,21 +1,12 @@
 import { Suspense, lazy } from "react";
-import { t } from "ttag";
 
 import { connect } from "metabase/redux";
 import type { State } from "metabase/redux/store";
 import { getSetting } from "metabase/settings";
 import TableSkeleton from "metabase/visualizations/components/skeletons/TableSkeleton";
-import {
-  getDefaultSize,
-  getMinSize,
-} from "metabase/visualizations/shared/utils/sizes";
-import type {
-  VisualizationDefinition,
-  VisualizationProps,
-} from "metabase/visualizations/types";
+import type { VisualizationProps } from "metabase/visualizations/types";
 
-import { _columnSettings as columnSettings, settings } from "./settings";
-import { checkRenderable, isSensible } from "./utils";
+import { PIVOT_TABLE_DEFINITION } from "./definition";
 
 const mapStateToProps = (state: State) => ({
   fontFamily: getSetting(state, "application-font"),
@@ -40,21 +31,7 @@ function PivotTableComponent(
   );
 }
 
-const PivotViz: VisualizationDefinition = {
-  getUiName: () => t`Pivot Table`,
-  identifier: "pivot",
-  iconName: "pivot_table",
-  minSize: getMinSize("pivot"),
-  defaultSize: getDefaultSize("pivot"),
-  canSavePng: false,
-  isSensible,
-  checkRenderable,
-  settings,
-  columnSettings,
-  isLiveResizable: () => false,
-};
-
 export const PivotTable = Object.assign(
   connect(mapStateToProps)(PivotTableComponent),
-  PivotViz,
+  PIVOT_TABLE_DEFINITION,
 );
